backend/requirements/base.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
django-storages==1.14.2
Pillow==10.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
django-celery-beat==2.6.0
channels==4.0.0
channels-redis==4.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.34.23
requests==2.31.0
